requirements.txt
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
filelock==3.16.1
fonttools==4.54.1
fsspec==2024.10.0
imageio==2.35.1
Jinja2==3.1.4
kiwisolver==1.4.7
MarkupSafe==3.0.2
matplotlib==3.9.2
mpmath==1.3.0
networkx==3.4.1
numpy>=1.26.0
packaging==24.1
pandas>=2.1.0
pillow==10.4.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
scipy==1.14.1
six==1.16.0
sympy==1.13.1
torchaudio>=2.2.0
torchvision>=0.17.0
tqdm==4.67.0
typing_extensions==4.12.2